requirements.txt
dj-database-url==0.4.2
Django==1.10.6
gunicorn==19.6.0
psycopg2==2.7
python-decouple==3.0
social-auth-app-django==1.1.0
whitenoise==3.3.0
mock==2.0.0
coveralls==1.2.0
docopt==0.6.2
Pillow==4.3.0
sorl-thumbnail==12.3